refactor: moved tests to a file
